requirements.txt
langchain
langchain_core
langchain_groq
langchain_community
python-dotenv
streamlit
langchain-google-genai